refactor: Move CustomTitleDialogStateModel location
