+++ feature/home/src/main/java/com/teamoffroad/feature/home/presentation/component/character/EmblemNameText.kt
@@ -7,7 +7,6 @@ import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.material3.IconButton
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
@@ -19,7 +18,7 @@ import com.teamoffroad.core.designsystem.component.OffroadTagItem
 import com.teamoffroad.core.designsystem.theme.OffroadTheme
 import com.teamoffroad.core.designsystem.theme.Sub
 import com.teamoffroad.core.designsystem.theme.White
-import com.teamoffroad.feature.home.presentation.component.dialog.CustomTitleDialogState
+import com.teamoffroad.feature.home.presentation.model.CustomTitleDialogStateModel
 import com.teamoffroad.offroad.feature.home.R
 
 @SuppressLint("UnrememberedMutableState")
@@ -27,7 +26,7 @@ import com.teamoffroad.offroad.feature.home.R
 fun EmblemNameText(
     modifier: Modifier = Modifier
 ) {
-    val customTitleDialogState = remember { mutableStateOf<CustomTitleDialogState?>(null) }
+    val customTitleDialogStateModel = remember { mutableStateOf<CustomTitleDialogStateModel?>(null) }
 
     val showDialog = mutableStateOf(false)
     var selectedCharacterIdx = remember { mutableStateOf<String?>("초보 모험가") }
@@ -60,10 +59,10 @@ fun EmblemNameText(
                 if (showDialog.value) {
                     OffroadDialog(
                         showDialog,
-                        customTitleDialogState,
+                        customTitleDialogStateModel,
                         onClickCancel = {
                             showDialog.value = false
-                            customTitleDialogState.value?.onClickCancel
+                            customTitleDialogStateModel.value?.onClickCancel
                         },
                         onCharacterChange = { idx ->
                             selectedCharacterIdx.value = idx.toString()
