+++ feature/home/src/main/java/com/teamoffroad/feature/home/presentation/model/CustomTitleDialogStateModel.kt
@@ -1,6 +1,6 @@
-package com.teamoffroad.feature.home.presentation.component.dialog
+package com.teamoffroad.feature.home.presentation.model
 
-data class CustomTitleDialogState( // 나중에 구조 보고 변경 예정
+data class CustomTitleDialogStateModel( // 나중에 구조 보고 변경 예정
     val title: String = "",
     val description: String = "",
     val onClickConfirm: () -> Unit = {},
